requirements.txt
marshmallow==3.14.0
marshmallow-oneofschema==3.0.1
click==8.0.3
GitPython==3.1.24